// extensions/quarto/packages/editor/src/api/command-types.ts
export interface EditorCommand {
  readonly id: EditorCommandId;
  readonly keymap: readonly string[];
  readonly isEnabled: () => boolean;
  readonly isActive: () => boolean;
  readonly plural: () => number;
  readonly execute: () => void;
}